Run through the full Photon tutorial
Now that testing works. need to be aware of a few things. We can  not be on the MIT Guest network and the sync of the players needs to ensure that everything is observed by photon view and that we make a seperation between the current players and the network players